chore: created posts router file
